fix: member/components 파일 구변경
