--- src/member/components/SignUpAddress/SignUpAddress.jsx
@@ -1,12 +1,0 @@
-export default function SignUpAddress() {
-  return (
-    <div>
-      <p>
-        <span>주소</span>
-        <span>*</span>
-      </p>
-      <button type="button">인증번호 받기</button>
-      <p>배송지에 따라 상품 정보가 달라질 수 있습니다.</p>
-    </div>
-  );
-}
