+++ src/member/components/SignUpAddress/index.jsx
@@ -1,3 +1,12 @@
-import SignUpAddress from './SignUpAddress';
-
-export default SignUpAddress;
+export default function SignUpAddress() {
+  return (
+    <div>
+      <p>
+        <span>주소</span>
+        <span>*</span>
+      </p>
+      <button type="button">인증번호 받기</button>
+      <p>배송지에 따라 상품 정보가 달라질 수 있습니다.</p>
+    </div>
+  );
+}
