--- src/member/components/SignUpBasicInput/SignUpBasicInput.jsx
@@ -1,20 +1,0 @@
-export default function SignUpBasicInput({
-  title, required, placeholder, name, pattern,
-  component,
-}) {
-  return (
-    <div>
-      <p>
-        <span>{title}</span>
-        {required ? <span>*</span> : null}
-      </p>
-      <input
-        name={name}
-        type="text"
-        placeholder={placeholder}
-        pattern={pattern}
-      />
-      {component}
-    </div>
-  );
-}
