+++ src/member/components/SignUpBasicInput/index.jsx
@@ -1,3 +1,20 @@
-import SignUpBasicInput from './SignUpBasicInput';
-
-export default SignUpBasicInput;
+export default function SignUpBasicInput({
+  title, required, placeholder, name, pattern,
+  component,
+}) {
+  return (
+    <div>
+      <p>
+        <span>{title}</span>
+        {required ? <span>*</span> : null}
+      </p>
+      <input
+        name={name}
+        type="text"
+        placeholder={placeholder}
+        pattern={pattern}
+      />
+      {component}
+    </div>
+  );
+}
